make dir & move files
